requirements.txt
anthropic>=0.40.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
numpy>=1.26.0
python-docx>=1.1.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
python-dotenv>=1.0.0
flask>=3.0.0
